Include button to update grid size
I have added a button to the page so the users can update the dimensions
of the grid.

Currently this just updates the number of blocks on the page but breaks
through the grid size.

The next iteration should adjust the block size so they fit in the same
grid despite having more squares.