Add logic to dynimcally update grid dimensions
Put logic in place within the Javascript file to update the size of each
box within the grid based on the user's input when clicking the button.

The grid also stays within the same container so it is just the box
sizes that change and the grid stays the same size.